user_study/user_study/backend/requirements.txt
aiofiles==24.1.0
boto3==1.35.48
botocore==1.35.48
dill==0.3.9
jose==1.0.0
pandas==2.2.3
passlib==1.7.4
Pillow==11.0.0
python-dotenv==1.0.1
scikit_learn==1.5.2
uvicorn==0.32.0
tensorflow==2.13.0
tensorflow-estimator==2.13.0
typing-extensions
tensorboard==2.13.0
keras==2.13.1
gast==0.4.0
fastapi==0.85.0
pydantic==1.10.0
sqlalchemy==1.4.45
torch==1.13.1
torchvision==0.14.1
numpy
python-jose
pydantic-core==0.42.0
pymysql
opencv-python
bcrypt